+++ packages/lib-content/src/components/AboutHeader/AboutHeader.stories.js
@@ -1,0 +1,8 @@
+import AboutHeader from './AboutHeader.js'
+
+export default {
+  title: 'Shared / AboutHeader',
+  component: AboutHeader
+}
+
+export const Default = {}