+++ packages/lib-react-components/CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Added
 - Added a `bodyBackground` prop to Modal to allow for custom background colors.
+- Added a `ProjectCard` component originally developed in lib-user, but now need for widespread FEM use.
 
 ### Changed
 - Removed `pxToRem` function.